latest_code

+++ models/official/projects/maskformer/tasks/panoptic_maskformer.py
@@ -294,7 +294,8 @@ class PanopticTask(base_task.Task):
 			# print("Output labels :", pred_labels)
 			print("Saving outputs...........", self.DATA_IDX)
 			# Save model inputs and outputs for visualization.
-			np.save("./input_img_"+str(self.DATA_IDX)+".npy", features.numpy())
+			# convert from bfloat16 to unit8 for image
+			np.save("./input_img_"+str(self.DATA_IDX)+".npy", tf.cast(features, dtype=tf.uint8).numpy())
 			np.save("./output_labels_"+str(self.DATA_IDX)+".npy", pred_labels.numpy())	
 			np.save("./target_labels_"+str(self.DATA_IDX)+".npy", labels["unique_ids"].numpy())
 			np.save("./output_masks_"+str(self.DATA_IDX)+".npy", outputs["mask_prob_predictions"].numpy())
